Armado server base, no recibe siempre los mensajes.

+++ server/entity.h
@@ -1,6 +1,8 @@
 #ifndef __ENTITY_H
 #define __ENTITY_H
 
+#include <cstdint>
+
 enum entity_type_t { PLAYER, NPC, LOOT };
 
 class Entity {
@@ -14,8 +16,9 @@ class Entity {
     /* Saber el tipo de la entidad sirve para saber si es colisionable o no.
     Los players y npcs son colisionables; el loot no.*/
     virtual entity_type_t get_type() = 0;
-    
-    virtual void update(unsigned int delta_t);
+
+    virtual void update(uint64_t delta_t);
+
     unsigned int get_id() const;
 };
